fix issue with proptypes and rendering of offers-list on offer page
