+++ src/components/main/main.jsx
@@ -6,15 +6,23 @@ import CitiesList from '../cities-list/cities-list.jsx';
 import Map from '../map/map.jsx';
 import OfferList from '../offers-list/offers-list.jsx';
 import NoOffers from '../no-offers/no-offers.jsx';
+import SortMenu from '../sort-menu/sort-menu.jsx';
+import withSortType from '../../hocs/with-sort-type/with-sort-type';
 
+const SortMenuWrapped = withSortType(SortMenu);
 
 const Main = (props) => {
-  const {offers} = props;
+  const {offers, activeCityName} = props;
 
   const renderOffersList = () => {
     return (
       <div className="cities__places-container container">
-        <OfferList/>
+        <section className="cities__places places">
+          <h2 className="visually-hidden">Places</h2>
+          <b className="places__found">{offers.length} places to stay in {activeCityName}</b>
+          <SortMenuWrapped/>
+          <OfferList/>
+        </section>
         <div className="cities__right-section">
           <section className="cities__map map">
             <Map/>
@@ -72,11 +80,13 @@ const Main = (props) => {
 };
 
 Main.propTypes = {
-  offers: PropTypes.arrayOf(offerPropType).isRequired
+  offers: PropTypes.arrayOf(offerPropType).isRequired,
+  activeCityName: PropTypes.string.isRequired
 };
 
 const mapStateToProps = (state) => ({
-  offers: state.offers.offers
+  offers: state.offers.offers,
+  activeCityName: state.offers.activeCityName
 });
 
 export default connect(mapStateToProps, null)(Main);
